Move logic to one file.
